+++ packages/ragbits-core/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Make the score in VectorStoreResult consistent (always bigger is better)
 - Add router option to LiteLLMEmbedder (#440)
 - Fix: make unflatten_dict symmetric to flatten_dict (#461)
+- Cost and capabilities config for custom litellm models (#481)
 
 ## 0.12.0 (2025-03-25)
 - Allow Prompt class to accept the asynchronous response_parser. Change the signature of parse_response method.