+++ packages/ragbits-core/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add router option to LiteLLMEmbedder (#440)
 - New methods in Prompt class for appending conversation history (#480)
 - Fix: make unflatten_dict symmetric to flatten_dict (#461)
+- Cost and capabilities config for custom litellm models (#481)
 
 ## 0.12.0 (2025-03-25)
 - Allow Prompt class to accept the asynchronous response_parser. Change the signature of parse_response method.