+++ packages/ragbits-core/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - Make the score in VectorStoreResult consistent (always bigger is better)
 - Add router option to LiteLLMEmbedder (#440)
+- New methods in Prompt class for appending conversation history (#480)
 - Fix: make unflatten_dict symmetric to flatten_dict (#461)
 - Cost and capabilities config for custom litellm models (#481)